+++ Source/Inspector.lua
@@ -27,7 +27,7 @@ function addonTable:InitialiseInspector()
         "GameFontNormalSmall")
     self.inspectorWindow.text:SetPoint("CENTER", self.inspectorWindow, "CENTER", 0, 10)
     self.inspectorWindow.text:SetText(
-        "Hover over any frame, and a tooltip will appear\n It will show you all the texts used by said frame")
+        "Hover over any frame, and a tooltip will appear\n It will show all font instances used by said frame")
     local inspectorHeader = CreateFrame("Frame", "FontmancerInspectorHeader", self.inspectorWindow,
         "DialogHeaderTemplate")
     inspectorHeader.text = inspectorHeader:CreateFontString("FontmancerInspectorHeaderText", "OVERLAY", "GameFontNormal")
@@ -45,33 +45,35 @@ function addonTable:InitialiseInspector()
     self.inspectorTooltip = CreateFrame("GameTooltip", "FontmancerInspectorTooltip", UIParent, "GameTooltipTemplate")
 end
 
-function addonTable:ConcatenateFontTables(firstTable, secondTable)
-    for _, value in pairs(secondTable) do
-        table.insert(firstTable, value)
-    end
-
-    return firstTable
-end
-
-function addonTable:GetInspectedFonts(frame)
+function addonTable:GetFontStrings(frame)
     local fontStrings = {}
+    local seen = {}
+
+    local function AddIfNotSeen(obj)
+        if obj and not seen[obj] then
+            seen[obj] = true
+            table.insert(fontStrings, obj)
+        end
+    end
 
     -- Check if the object itself is a font string
     local isPossiblyFontString, isFontString = pcall(function() return frame:IsObjectType("FontString") end)
     if isPossiblyFontString and isFontString then
-        table.insert(fontStrings, frame)
+        AddIfNotSeen(frame)
     end
 
     -- Check if the object has a direct font string
     if frame.GetFontString and frame:GetFontString() then
-        table.insert(fontStrings, frame:GetFontString())
+        AddIfNotSeen(frame:GetFontString())
     end
 
     -- Check if the object's regions is/has a font string
     local hasPossiblyRegions, regions = pcall(function() return { frame:GetRegions() } end)
     if hasPossiblyRegions then
         for _, region in pairs(regions) do
-            fontStrings = self:ConcatenateFontTables(fontStrings, self:GetInspectedFonts(region))
+            for _, fontObjects in ipairs(self:GetFontStrings(region)) do
+                AddIfNotSeen(fontObjects)
+            end
         end
     end
 
@@ -106,17 +108,17 @@ function addonTable:GetFocusedFrames()
     end
 
     -- Nameplates
-    local function TryAddFrame(frame)
-        if not frame:IsForbidden() then
-            table.insert(focusedFrames, frame)
+    local function TryAddNameplate(nameplate)
+        if not nameplate:IsForbidden() then
+            table.insert(focusedFrames, nameplate)
         end
     end
 
     if C_NamePlate and C_NamePlate.GetNamePlateForUnit then
         local nameplate = C_NamePlate.GetNamePlateForUnit("mouseover")
         if nameplate then
-            TryAddFrame(nameplate)
-            TryAddFrame(nameplate.UnitFrame)
+            TryAddNameplate(nameplate)
+            TryAddNameplate(nameplate.UnitFrame)
         end
     end
 
@@ -126,32 +128,27 @@ end
 function addonTable:PopulateTooltip()
     self:ClearTooltip()
 
-    local frames = self:GetFocusedFrames()
+    local focusedFrames = self:GetFocusedFrames()
 
     -- Sort by visual stacking order
-    table.sort(frames, function(a, b)
+    table.sort(focusedFrames, function(a, b)
         return (a:GetFrameLevel() or 0) > (b:GetFrameLevel() or 0)
     end)
 
-    for _, focusedFrame in pairs(frames) do
+    for _, focusedFrame in pairs(focusedFrames) do
         if focusedFrame ~= WorldFrame and focusedFrame ~= UIParent then
-            local fontStrings = self:GetInspectedFonts(focusedFrame)
+            local fontStrings = self:GetFontStrings(focusedFrame)
 
             if #fontStrings > 0 then
                 local headerColour = self.databaseDefaults.colours.text
                 self.inspectorTooltip:AddLine(focusedFrame:GetDebugName(), headerColour.r, headerColour.g, headerColour
                     .b)
-                self.inspectorTooltip:AddLine(" Uses font objects:")
 
                 for _, fontString in pairs(fontStrings) do
-                    local regionName = fontString:GetName() or "Anonymous"
-                    local fontObject = fontString:GetFontObject()
-                    local fontInstanceName = fontObject and fontObject:GetName()
-                    if fontInstanceName then
-                        self.inspectorTooltip:AddLine("  - " .. regionName .. ": |cffffffff" .. fontInstanceName .. "|r")
-                    else
-                        self.inspectorTooltip:AddLine("  - " .. regionName .. ": |cff808080(No Global Instance)|r")
-                    end
+                    local fontStringName = fontString:GetName() or fontString:GetDebugName() or "Anonymous"
+                    local fontInstanceName = fontString:GetFontObject():GetName()
+                    self.inspectorTooltip:AddLine("  - " ..
+                        fontStringName .. ": |cffffffff" .. fontInstanceName .. "|r")
                 end
                 self.inspectorTooltip:AddLine(" ")
             end